requirements.txt
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.4
scipy==1.11.4
PyYAML==6.0.1
werkzeug==3.0.1
tqdm==4.66.1
requests==2.31.0
psutil==5.9.6
